Fix warnings and cast issues. Compiles with -Werror now.
Deprecated char* casts are a lost case. -Wno-write-strings ...

+++ src/burn/sek.cpp
@@ -1417,7 +1417,7 @@ int SekMapMemory(unsigned char* pMemory, unsigned int nStart, unsigned int nEnd,
 	return 0;
 }
 
-int SekMapHandler(unsigned int nHandler, unsigned int nStart, unsigned int nEnd, int nType)
+int SekMapHandler(uintptr_t nHandler, uintptr_t nStart, uintptr_t nEnd, int nType)
 {
 	unsigned char** pMemMap = pSekExt->MemMap + (nStart >> SEK_SHIFT);
